[Workflows] Cancel form-filling animation when Print/Download button is selected (Recidiviz/recidiviz-dashboards#2425)
This PR fixes the bug in which selecting Print PDF before the form-filling animation is complete leads to the downloaded PDF being complete. This change introduces a side effect based on the formIsPrinting observable on the selected client, which cancels the animation by setting mountedAt to 0 and animated to true so that the full value is filled in.

GitOrigin-RevId: 597aa92c7fd006d16372935e110ac06e645eded5