[Workflows] Implement specific sorting by opportunity (Recidiviz/recidiviz-dashboards#2465)
This PR instantiates a mapping between opportunity type to sort function. This enables us to specify how to sort a specific opportunity. This PR also introduces a new Date field on the OpportunityBase class that will be used to specify an urgency date to act on the opportunity, which will be used for the new opportunities in Idaho.

GitOrigin-RevId: adec2fbd3cc13e8b0024e9909b369d3d8fafe143