[Workflows] feat: Add feature variants for US_Id tenant (Recidiviz/recidiviz-dashboards#3822)
* feat: Add feature variants for US Identity tenant

This commit adds two new feature variants, "usIdCRC" and "usIdExtended", for the US Identity tenant. These variants are used to control access to certain features based on the user's role. The commit also includes updates to the workflows store to handle the disallowed systems based on the tenant and role.

* feat: Update feature variant "usIdExtended" to "usIdExtendedCRC"

- Updated the feature variant "usIdExtended" to "usIdExtendedCRC" in the RootStore/types.ts file.
- Added a new property "featureVariantSupportedSystems" to the WorkflowsStore.ts file.
- Configured the support for the "usIdExtendedCRC" variant to the US_ID tenant in the US_ID.ts file.
- Added the "workflowsSupportedSystemsByFeatureVariant" property to the TenantConfig type in the core/models/types.ts file.

* feat: Add support for getting unique supported systems

This commit adds a new function to the WorkflowsStore that retrieves the list of supported systems based on the user's permissions. The function now returns a list of unique supported systems by using the `uniqWith` function from the lodash library.

* Change supported systems retrieval in WorkflowsStore

The code change modifies the retrieval of supported systems in the WorkflowsStore class. It replaces the variable name "result" with "supportedSystems" for clarity and consistency. This change improves readability and maintainability of the code.

* feat: Add feature variant mappings and gate workflows by feature variants

This commit adds feature variant mappings to the `RootStore/types.ts` file and updates the `defaultFeatureVariantsActive` constant to use the mappings. It also adds a new constant `allFeatureVariants` to store all feature variants. In the `WorkflowsStore/WorkflowsStore.ts` file, the getter `unsupportedWorkflowSystemsByFeatureVariants` is added to retrieve the unsupported workflow systems based on the feature variants. The `workflowsSupportedSystemsByFeatureVariant` property in the `core/models/types.ts` file is renamed to `workflowsGatedSystemsByFeatureVariant` and modified to store the allowed feature variants for each system. Additionally, in the `WorkflowsStore/__tests__/WorkflowsStore.test.ts` file, the test cases are updated to include the new feature variants `usIdCRC` and `usIdExtendedCRC` for the US_ID tenant.

* feat: Add support for US_ID workflows

- Added US_ID workflows to the WorkflowsStore
- Updated available state codes in the tenant configuration
- Added tests for opportunityTypes of US_ID

* Refactor lodash imports in WorkflowsStore.ts

This commit refactors the lodash imports in WorkflowsStore.ts to remove unused functions and improve code readability.

* feat: Improve handling of unsupported systems

For each `SystemId` that is gated by one or more `FeatureVariant`s, check if the user has any `FeatureVariant`s that allow the `SystemId`. If not, mark the `SystemId` as unsupported. This change improves the handling of unsupported systems by capturing all `SystemId`s that the user does not have permission to access.

In addition, update the test cases for opportunity types in the `US_BB` state.

* Add lodash's `difference` function and write tests for unaffected systems in `US_BB`.

- Imported the `difference` function from lodash to the test file.
- Added tests to check for unaffected systems in `US_BB` based on certain conditions.
- Updated expected values for tests on supported and unsupported systems.

* Did not make original push

* Fix bug in test case description for workflowsStore.workflowsSupportedSystems

* Refactor state code check in WorkflowsStore.test.ts

- Update the state code check in the test to use a constant variable instead of a hard-coded value.
- Change the detailed message in the test to reflect the new state code variable.

GitOrigin-RevId: 518cafbcb4a104425107aef3d562714b5fafba42