[Workflows][US_TN] Create `UsTnAnnualReclassificationReferralRecord` (Recidiviz/recidiviz-dashboards#4040)
* feat: Add UsTnReclassificationReviewReferralRecord and config file

This commit adds the UsTnReclassificationReviewReferralRecord definition and its associated config file for the Opportunity module in the UsTn workflow. This allows for the handling and validation of reclassification review referral records specific to Tennessee.

* feat(FirestoreStore): add new collection name 'usTnAnnualReclassificationReferrals'

This commit adds a new collection name 'usTnAnnualReclassificationReferrals' in the FirestoreStore.

Rename files

This commit renames files related to the 'UsTnReclassificationReviewOpportunity' to 'UsTnAnnualReclassificationReviewOpportunity', including the schema, config, and index files.

Add test for UsTnAnnualReclassificationReviewReferralRecordRaw

This commit adds a test for the 'UsTnAnnualReclassificationReviewReferralRecordRaw' type, ensuring that it can be constructed and matches the snapshot.

* Refactor schema to allow optional and nullable values

The schema for `UsTnAnnualReclassificationReviewReferralRecord` has been modified to allow the `custodyLevelNotMax` field to be optional and the `custodyLevelComparedToRecommended` field to have optional and nullable string values.

Additionally, a new schema `usTnAnnualReclassificationReviewSchema` has been added to `testZodSchema.ts` to handle reclassification referrals in Tennessee.

* change in config names used

* Update config.ts

change to annual reclassifcation name

* fix

* change from nullable to optional

* refactor for OpportunitySchemaBase

* feat: Add caseNotes to UsTnAnnualReclassificationReviewReferralRecord

- Added `caseNotesSchema` to `usTnAnnualReclassificationReviewSchema` in `UsTnAnnualReclassificationReviewReferralRecord.ts`
- Updated tests in `UsTnAnnualReclassificationReferralRecord.test.ts` to include caseNotes
- Added a snapshot test for the parsed data without caseNotes in `UsTnAnnualReclassificationReviewReferralRecordRaw`

* Update UsTnAnnualReclassificationReferralRecord.test.ts.snap

- Add caseNotes to the snapshot
- Update the stateCode from "US_XX" to "US_TN"

This commit updates the snapshot for the UsTnAnnualReclassificationReferralRecord test. It adds caseNotes to the snapshot and updates the stateCode to "US_TN" from "US_XX".

* replace criteria to become custodyLevel

GitOrigin-RevId: 5ae7511009ab6ea2bbab5db963c8be802d9a852d