[US_TN][OppConfig][Workflows]  Move `generateopportunityinitialheader()` (Recidiviz/recidiviz-dashboards#4089)
* feat: add initialHeaders to opportunity configs

- Updated the OpportunityConfigs.ts file to include an optional property called "initialHeader" in the OpportunityConfig type.
- Added the "initialHeader" property to the UsIdPastFTRDOpportunity, UsMeEarlyTerminationOpportunity, UsMeSCCPOpportunity, and UsTnExpirationOpportunity configs.
- Modified the utils.ts file to use the "initialHeader" property when generating the opportunity initial header.
- If the opportunity config does not have an "initialHeader" property, the utils.ts file generates the header based on the opportunity type, justice involved person title, and workflows search field title.

* Refactor OpportunityConfigMap type

The OpportunityConfigMap type in OpportunityConfigs.ts file has been changed to be exported. This change allows other modules to import and use the type. This change was made to improve code reusability and maintainability.

Add opportunity configuration fixture

A new opportunity configuration fixture has been added to the __fixtures__/index.ts file. The fixture includes a mock opportunity configuration with various properties such as stateCode, urlSection, label, featureVariant, initialHeader, customTabOrder, and snooze. This fixture is used for testing purposes.

Add opportunity configuration map fixture

A new opportunity configuration map fixture has been added to the __fixtures__/index.ts file. The map includes the existing OPPORTUNITY_CONFIGS and the newly added mock opportunity configuration. This map is used for testing purposes.

Add snapshots for utility functions

Snapshots have been added for the generateOpportunityInitialHeader function in the __tests__/__snapshots__/utils.test.ts.snap file. The snapshots ensure that the expected output of the function remains unchanged.

Update utility function tests

The utility functions in the utils.test.ts file have been updated to use the new opportunity configuration fixtures. The tests now include cases for generating the opportunity initial header when the initialHeader property is provided in the configuration and when it is not provided. The tests ensure that the correct headers are generated based on the configuration.

This commit improves code reusability by exporting the OpportunityConfigMap type and adds fixtures and tests for the opportunity configuration and generation of opportunity initial headers.

* refactor var name from TEST_PERSON to TEST_TITLE

* Add mockUsXxOppConfig fixture and refactor header generation test in utils test file

- Refactored the header generation test in the utils.test.ts file to utilize a new fixture called mockUsXxOppConfig.
- Added the necessary import statement for the mockUsXxOppConfig fixture.
- Updated the test case to handle the initialHeader configuration provided in the config.

GitOrigin-RevId: 693e75a435d8a8bbb4b6cadcba74fe154bd1a045