[Workflows][OppConfigs] Generate headers from `OpportunityConfigs` (Recidiviz/recidiviz-dashboards#4114)
* partially moved hydrated headers

* Fix the types of `OpportunityHydratedHeader` and `generateOpportunityHydratedHeader`. defined in `OpportunityConfigs.ts`.

* feat: Add mockUsXxTwoOppConfig with hydrated header generation

This commit adds a new opportunity configuration `mockUsXxTwoOppConfig` that extends the existing `mockUsXxOppConfig` with a hydrated header generation function. The hydrated header includes a full text, opportunity text, and a call to action. Additionally, the commit updates the `MOCK_OPPORTUNITY_CONFIGS` map to include the new configuration.

refactor: Update utils.test.ts

This commit refactors the `utils.test.ts` file by importing the `generateOpportunityHydratedHeader` function and `OpportunityHydratedHeader` type. It also adds a new test suite to generate the hydrated header for each opportunity configuration in the `MOCK_OPPORTUNITY_CONFIGS` map. The generated hydrated headers are stored in the `hydratedHeaders` array for further testing. Finally, the commit includes tests to validate the generated hydrated headers and match them against the snapshot.

GitOrigin-RevId: 64116bcacd317ffab16df84dbc9c986ca76d53dd