[Workflows] Configure Firebase to Automate Backups (Recidiviz/recidiviz-dashboards#4223)
* Init firebase functions

* gitignore env

* set root eslint

* Add "type" field to package.json

- The "type" field with the value of "commonjs" has been added to the package.json file.

* update package lock

* The function now access google cloud secrets to get output dir

* "Initiate configuration for Firestore export"

This commit initiates the configuration process for exporting collections from Firestore. A configuration message is logged and displayed, showing the collections that will be exported from the specified project to the output bucket. The export is scheduled to run every Friday at 11:59 PM PDT.

* change logging for style

* var name fix

* another style change

* remove default from list of projects in `.firebaserc`

* removing logging changes--original impl was correct.

* refactor: Simplify code and improve error handling in export function

This commit simplifies the code by removing unnecessary imports and comments. It also improves error handling by throwing a custom error with a specific error message when no collections are specified.

The error message now includes the project ID and the output directory for better clarity. Additionally, a new file `utils.ts` is added to handle hold util functions

* create index, etc.

* remove secrets

* change readme

* readme edit

* readme fix

* fix

* gitignore edit

* readme fix

* read me fix

* Refactor configuration and export function in index.ts

- Refactor the configuration code to improve readability and remove unnecessary variables.
- Update the export function to use the correct output bucket.
- Improve error handling and logging.
- Provide more detailed success information in the logs.

GitOrigin-RevId: beca7ee6ee3610d0a38a80a4ff3333917d7a9c57