[Workflows][Configs] Refactor the Incarceration and Supervision typing (Recidiviz/recidiviz-dashboards#4181)
* feat: Add OpportunityConfigs and OpportunityConfig types

This commit adds new files for managing opportunity configurations and defining the types associated with them.

* feat: Add functions to check if a value is an incarceration or supervision opportunity type

This commit adds two new functions in the `OpportunityConfigs` module: `isIncarcerationOpportunityType` and `isSupervisionOpportunityType`. These functions can be used to check if a given value is an incarceration or supervision opportunity type respectively. The functions are implemented using a set of keys from the `OPPORTUNITY_CONFIGS` object and improve the code readability and reusability.

* Remove unnecessary opportunity types and update opportunity type mapping in types.ts

The commit message is a concise summary of the code changes.

* feat: Add opportunity config type with `T` for subsequent PR

Adds an opportunity config type with a generic `T` parameter to assert and type the constructor being used for the config. This will be used in a subsequent PR.

* Add opportunity configuration objects for different states

- Declares and exports opportunity configuration objects for states such as US_CA, US_ID, US_ME, etc.
- Each object contains specific configurations for that state.
- The objects are then merged into the `OPPORTUNITY_CONFIGS` map.

* simplify typing

* simplify typing

* feat(OPPORTUNITY_CONFIGS): Add state-specific configurations for incarceration and supervision opportunities

This commit adds state-specific configuration objects for incarceration and supervision opportunities. Each state has separate configuration objects for both types of opportunities. These configuration objects are aggregated into `INCARCERATION_OPPORTUNITY_CONFIGS` and `SUPERVISION_OPPORTUNITY_CONFIGS` respectively. Additionally, there are separate configuration objects for each state that aggregate both incarceration and supervision opportunities. The overall aggregation of all state opportunities is stored in `OPPORTUNITY_CONFIGS`. This commit also introduces utility functions `isIncarcerationOpportunityType()` and `isSupervisionOpportunityType()` to check if a given string represents an incarceration or supervision opportunity.

Please note that this commit also includes eslint-disable comments for unused imports.

* feat: Add core typing section for system stability.

* force save

* force save

* Refactor import statements for `SupervisionOpportunityType`

* Add opportunity workflow constants

This commit adds the constants file for the opportunity workflow in the WorkflowsStore directory. The constants file includes the necessary licensing information and specifies the terms under which the program is distributed.

* refactored how supervisionOpp and incarcerationOpp typing is generated, while also removing one mapping of types

* updated test config

* Delete unnecessary license comments in `constants.ts`

The commit removes license boilerplate comments from `constants.ts` in the `src/WorkflowsStore/Opportunity` directory. This change cleans up the code by removing unnecessary comments related to licensing.

* unit tests

* lint fix

* feat: Add system type to OpportunityConfigs (Recidiviz/recidiviz-dashboards#4293)

- Added  property to  type
- Modified existing opportunity configurations to include  property indicating the type of system (SUPERVISION or INCARCERATION) the opportunity belongs to
- Created helper functions  and  to retrieve the list of opportunity types specific to each system type

Co-authored-by: Daniela Schiano di Cola <daniela@recidiviz.org>
Co-authored-by: Daniel McLaughlin <danielm@recidiviz.org>

* Refactor OPPORTUNITY_CONFIGS to MOCK_OPPORTUNITY_CONFIGS

The OPPORTUNITY_CONFIGS constant is refactored to MOCK_OPPORTUNITY_CONFIGS in order to improve code readability and maintainability.

* updated tests and typing

* Refactor OpportunityConfig types in OpportunityConfigs.ts.

The OpportunityConfig type in OpportunityConfigs.ts has been modified to include a StateCode type parameter. This StateCode type parameter is used to represent the state code associated with an opportunity variant.

- The OpportunityConfig type now includes a new type parameter StateCode, which represents the state code associated with an opportunity variant.
- The stateCode field in the OpportunityConfig type has been updated to use the StateCode type parameter.
- The SupervisionOpportunityType and IncarcerationOpportunityType types have been updated to include the StateCode type parameter in the OpportunityConfig type.
- The OPPORTUNITY_CONFIGS_BY_STATE constant has been updated to use the StateCode type parameter to map each state code to its corresponding set of opportunity configurations.
- The OPPORTUNITY_TYPE_FOR_URL_BY_STATE constant has been updated to use the StateCode type parameter to map each state code to a mapping of URL sections to opportunity types.

This change allows for better type safety and encapsulation of the state code associated with each opportunity variant in the OpportunityConfigs module.

* leftover

* loosened state typing on `OpportunityConfig`s

* Refactor opportunity configuration mapping for improved readability and type safety.

- Extracted the generic `getOpportunityTypesBySystemType` function to support a wider range of opportunity types.
- Adjusted the type signature of `getOpportunityTypesBySystemType` to work with the updated opportunity configuration structure.
- Simplified the `ConfigsByStateMapping` type to use `Record` for better readability.
- Updated `OPPORTUNITY_CONFIGS_BY_STATE` to utilize the updated `ConfigsByStateMapping` type.

---------

Co-authored-by: Daniela Schiano di Cola <daniela@recidiviz.org>
Co-authored-by: Daniel McLaughlin <danielm@recidiviz.org>
GitOrigin-RevId: 2318d32a52fd8cb9374e6b00a26698da6838f1e4