[Workflows][US_ME] Update forms and create a download flow that downloads multiple file formats (Recidiviz/recidiviz-dashboards#4335)
* fix: check

* [DOCXFormGenerator]: Add new functions for rendering and downloading multiple DOCX files.

This commit adds two new functions, `renderDocx` and `renderMultipleDocx`, to the `DOCXFormGenerator` module. The `renderDocx` function renders a single DOCX document based on the provided template and form contents. The `renderMultipleDocx` function renders multiple DOCX documents at once based on an array of input data.

Additionally, the commit introduces a new module `Zipper.ts` for handling the creation and download of ZIP files. The `downloadZipFile` function in this module takes an array of file objects with filenames and file contents, and creates a ZIP file containing all the files. The resulting ZIP file is then downloaded to the user's device.

These changes enable the downloading of multiple DOCX files in a single ZIP file, providing a more efficient and convenient way to distribute generated documents.

* Fix file saving issues and add error capture

The commit introduces changes to fix file saving issues and adds error capture for better error handling. It also updates the use of external libraries for creating and saving zip files.

* FIX

* [Paperwork] Move downloadZipFile function to utils

The downloadZipFile function was moved from the Zipper file to the utils file. This change was made to consolidate similar functionality in one location. The function is used to download a zip file containing multiple files.

* lint fix

* Fix file name capitalization and refactor form downloading function in FormFurloughRelease component

The commit renames the file names in the server directory to have consistent capitalization. It also improves the `FormFurloughRelease` component by refactoring the `FormDownloader` function to `formDownloader` and updating the file name formatting logic using `fileNameFormatter`.

GitOrigin-RevId: 0a521084e98be4e706f262e91af127a26886a912