[Revamp cleanup] remove ResponsiveRevamp profile type (Recidiviz/recidiviz-dashboards#4846)
Recidiviz/recidiviz-dashboards#4614

PR 3/x to remove the responsiveRevamp feature variant.

This PR:

removes the ResponsiveRevamp type which was used in a bunch of intersection props types in the FullProfile
removes the responsiveRevamp prop from all the components that were passing it through, and addresses the obsolete logic at the leaf components.

GitOrigin-RevId: cc85c90c02fdf231e82e77b80b6af3cc57a21c1b