[Revamp cleanup] Remove feature flag in OpportunityCaseloadView and its dependencies (Recidiviz/recidiviz-dashboards#4856)
Recidiviz/recidiviz-dashboards#4614
PR 4/x to remove the responsiveRevamp feature variant. I'm expecting the remaining to changes to be similar in that I'll chunk the work by directories where I can.

This PR:

handles the removal of the feature flag in the OpportunityCaseloadView directory and its dependencies
In particular, OpportunityPersonList, OpportunityPreviewModal were updated, and then whatever changes were needed for the relevant tests to pass.
Removes OpportunityPersonListWithSectionTitles as it's no longer used anywhere (previously referenced in OpportunityPersonList.tsx).

GitOrigin-RevId: 723119791569110ec45058ed45827d754d911ced