[Revamp cleanup] remove flag from Caseload directories (Recidiviz/recidiviz-dashboards#4888)
Recidiviz/recidiviz-dashboards#4614

This PR:

handles the removal of the feature flag in the CaseloadSelect, CaseloadView, and CaseloadTypeSelect directories.
does a bit more work in CaseloadSelect.tsx to trim anything unused. The GH line diff came out a bit wonky in some places but the gist of larger changes is:
combines the baseStyles object with newStyles, and renames to just styles since there's no need to assign an object based on the feature flag anymore
removed an obsolete Disabled menu list component.

GitOrigin-RevId: 2d22c5feeaa74c0f668b5478ccf01d98443f78e7