Add record schema and snapshot test for schema parsing for PA AS (Recidiviz/recidiviz-dashboards#4929)
Initial ZOD schema for the admin supervision record in usPaAdminSupervisionReferralRecord.ts (only eligible criteria for now). Adds a couple unit tests for parsing verification (+ snapshots generated). Also adds a fixture to the US_PA index.ts fixture file which is not yet in use, but perhaps for future opportunity testing it will be.

GitOrigin-RevId: 11585307131dad05d82f5c5f1b44adfa86e4714c