Pull birthdate into OR ED form on FE (Recidiviz/recidiviz-dashboards#5260)
This PR finishes up the work to start populating the AOS birthdate on the OR ED forms.

Updates the ZOD schema and form data type to include the birthdate. Also updates snapshot and fixture files. Tested via offline mode and then dev mode (confirmed the populated birthdate for a few clients matched the data in the us_or_earned_discharge_record_materialized table in bq).

GitOrigin-RevId: 1a0dff55150a17d07a9f36be897f73af64863123