Remove criteria on FE for specific supervision level (Recidiviz/recidiviz-dashboards#5265)
Recidiviz/recidiviz-dashboards#5253

For PA admin supervision, on the FE we previously had displayed the supervisionLevelIsNotLimited eligibility criteria using the bullet point "Currently on [x] supervision". However this isn't necessary, as folks don't need to be on a specific supervision level to be eligible, they just can't already be on administrative supervision (which we check for in the relevant eligibility spans view). This PR removes the supervisionLevelIsNotLimited as a specific eligibility criteria on the FE.

GitOrigin-RevId: 0ea71cccb84ddc6b60a34bc7a2a65cdb70fd600b