[Workflows][US_MI]Adds first three SCC reviewform sections and auto-fills from data (Recidiviz/recidiviz-dashboards#5375)
Adds initial three sections to the MI SCC review form component (previously stubbed). This is a pretty big form, so there's going to be a couple more PRs to follow.

Also, separates out a (hopefully) generic DOCXCheckbox component, building off of the PA form checkbox, and the generic DOCX input field component. I didn't replace existing checkbox implementations with this one, but ideally can be used in the future.

GitOrigin-RevId: aeba3f5569a5245a1fd965352f6052ebeaffd946