[Workflows][Single-instance forms] Move draft data update function to FormBase (Recidiviz/recidiviz-dashboards#5483)
Consolidating form update logic into the FormBase as part of the work to switch over to single-instance forms.

Tested inputs locally + confirmed draft data updated/cleared in firestore emulator.
Added unit tests as well.

In the next ticket, we'll start reading/writing updates to the new collection - at which point I believe we'll want to add an async updateForm fn or something similar to the FirestoreStore and start using that behind the FV

GitOrigin-RevId: 3c56b49637f0fae848b0f5496ef790bb59e7b021