[Workflows][US_MI][Bug bash] Update form template, tab order, and some copy (Recidiviz/recidiviz-dashboards#5545)
Now that single instance form functionality is in prod, we should update the FV to be on by default for recidiviz users. Also addresses some various bug bash tickets as follows.

Recidiviz/recidiviz-dashboards#5522: In the form template, the second "Regularly" checkbox was labeled regularly1 instead of regularly2, so when we downloaded the form, it was showing the incorrect value.
Recidiviz/recidiviz-dashboards#5517: Form template was missing the date field for the various segregation types.
Recidiviz/recidiviz-dashboards#5524: Adds an almostEligibleStatusMessage ("Upcoming") to the SCC opps to match the tab name.
Recidiviz/recidiviz-dashboards#5525: updates the tabOrder for SCC opps to order by action-most-needed

GitOrigin-RevId: 7c4e39b2c349ab24f1fbfff1e646957a6dfa537d