Update UI to use new swarm presenter and adds hover effect (Recidiviz/recidiviz-dashboards#5591)
PR 3/3 to finish hooking up v1 of supervisor homepage. Specifically, we need to support multiple officers highlighted on the same dot plot.

This creates a SwarmPlotContainerV2 that passes the SwarmPlotV2 the new swarm presenter instead of the legacy version (figured we should just make a v2 here since the props are branching). It also adds the hover effect from the designs to highlight officers in the sidepanel when their dot is hovered on the swarm plot.

Most of the changes are swapping in the V2 container into the relevant components. There's also some cleanup of the workaround swarmPlotMetric in the staff card and the legacy SwarmPresenter.

GitOrigin-RevId: 608fb92b315162911d09bbd13bb4fbc8c83243ae