[Supervisor homepage] Modify metric events table to display ID column (Recidiviz/recidiviz-dashboards#5642)
Some initial changes to modify the metric events table for the new UI. The figma designs are built on the "Small Desktop" breakpoint, so initially the ID column was always hidden because the events table was a bit skinny. This change gives the events table a bit more space on the page and also pulls some width away from the Name column.

The new design is here. There will need to be a bit more work done to match that UI exactly (modify header styling, column widths, and potentially give the table even more space relative to the plots), but it's P1.

GitOrigin-RevId: 9e87706404e31a92ef92f943c537f2c74c59809a