[Supervisor homepage] Add excluded officers endpoint to insights API and supervisor presenter (Recidiviz/recidiviz-dashboards#5737)
We recently updated the insights querier + routes to add an endpoint for a given supervisor's excluded officers (excluded officers are those not used to calculate outcomes/targets). This PR adds the endpoint to the insights API and the relevant presenter. It also includes a UI change on the SupervisorPageV2 to match the designs and pull in the excluded officers.

GitOrigin-RevId: 312225da7686be6cbb5f3a6608005515d880a61c