[Supervisor Homepage] modifications to the SupervisionOfficerPresente… (Recidiviz/recidiviz-dashboards#5798)
Pulls in workflows client and opp data in the shape we need using the JII store. See Recidiviz/recidiviz-dashboards#5820 for the workflows component that relies on these fields.

Ran in to some testing troubles with needing the opportunities to be fully hydrated state before the fns being tested would populate. Ended up just spying on/overriding the hydration state for the client's opps.

GitOrigin-RevId: 427c16c05f8bf5f48b47e7987c790a33ffdae9d5