[Supervisor homepage] Refactor OpportunitySummaries to be shared between insights and workflows (Recidiviz/recidiviz-dashboards#5820)
Refactors the OpportunitySummaries component so that the data needed for the UI can either come from the workflowsStore or from the SupervisionOfficerPresenter which is what we use for the supervisor homepage. The parent component handles the data acquisition. Then, pops the OppSummaries component into place on the officer view for the supervisor homepage and handles any other UI tweaks to match the figma.

There is a TODO(Recidiviz/recidiviz-dashboards#5320) for handling the "View all" link navigation since we don't yet have the page/URL set up for that on the supervisor homepage side.

GitOrigin-RevId: 409546ebaa837558df8347755037569f04d9fd96