[Workflows][single-instance forms] Update form classes to use new collection (behind FV) (Recidiviz/recidiviz-dashboards#5662)
This PR moves the isolateFormUpdates feature variant check to the FormBase and also has FormBase handle the creation of the formId field (using a formType from each child class and a class field shareFormUpdates that demarcates whether it's a shared form). When this lands, any users who have the isolateFormUpdates variant set (rn that's just Recidiviz folks and MI facilities staff) will begin using the new collection.

Tested forms in offline mode to confirm edits are going to/pulled from the new collection.

GitOrigin-RevId: 906bc1c69d9174167f0595ec0d09189f6dc27ac8