[Supervisor Homepage] Add new route for opportunity drill-down and link to page from summaries (Recidiviz/recidiviz-dashboards#5872)
Adds a new insights route for the opportunity page (/supervision/staff/:officerPseudoId/opportunity/:opportunityType) which will house the shared workflows OpportunityPersonList component. Updates the OpportunitySummaries to allow us to navigate to this new route when we're looking at the summaries from insights.

This also updates the InsightsPageLayout to allow for the page title to be optional (pulls out the title contents logic to make that easier). When we pull in the OpportunityPersonList component, we'll use that components title/subtitle rather than the insights page layout title. See this slack convo for details on that design decision.

First time adding an insights route so there were some uncertainties around params in the supervision store/permissions. Added comments for those.

GitOrigin-RevId: 21a9e7220c372b8772b93cf4ea26fa0aaae75dfc