[Supervisor homepage] Initial changes to OpportunityPersonList to prepare for shared refactor (Recidiviz/recidiviz-dashboards#5897)
This PR:

Adds opportunityType as a required prop for the HydratedOppPersonList and then removes any relevant conditional logic in the component.
allOpportunitiesByType on the workflowsStore by definition will always exist (thought it may be an empty obj), so removes conditional logic around that.
Also removes a test which mocks this value to be undefined.
The empty state is handled by the CaseloadOpportunitiesHydrator so removes the additional empty state from the HydratedOppPersonList (see hydrator and hasOpportunities logic; it covers the case were the array length is 0 for a particular opportunityType)
updates a test to check empty state when hasOpportunities returns false.

GitOrigin-RevId: 2fa27ace1371cc90e5442a5c0b2dc18607f4e7a1