[Workflows] Use the correct searchField when defining the query for the resident CaseloadSubscription (Recidiviz/recidiviz-dashboards#5946)
We have to specifically handle activeSystem === "ALL" when setting up a location-based query. This PR moves the logic for pulling the locationSearchField from the LocationSubscription into a WorkflowsStore getter. Then updates the CaseloadSubscription to use that field instead of the default searchField when we are in "ALL" mode and it's a resident subscription.

GitOrigin-RevId: 610aafdce4276c0c8f476349fcb51a8240871072