[Workflows] Add deletion to migration script for form updates (Recidiviz/recidiviz-dashboards#5940)
Final cleanup step for the form updates migration. We can fully remove the referralForm field from clientOpportunityUpdates docs once all forms and users have switched over to using the clientFormUpdates.

Should wait to run this til it's clear we won't need to roll back to the old version.

Tested in offline mode and did a dry run in staging.

GitOrigin-RevId: 11d9fedf923a82c2ad601afe75e0cd86391ad785