[Supervisor Homepage] Share the HydratedOpportunityList component between workflows and insights (Recidiviz/recidiviz-dashboards#5895)
This PR:

Moves the HydratedOpportunityList into it's own file
Moves the larger data structures used in the component to be props, which can then be passed in from either the supervisor homepage presenter or the workflows store.
Still uses a wrapper component on the workflows side for the workflowsStore data structures because there was an earlier refactor to help with load time and I don't want these changes to roll that back.
Moves the opportunitiesByTab helper function out of the workflowsStore (since it doesn't depend on any store fields) and into the utils file.
updates the relevant test to handle mocking this change
Handles the selectedPerson field when accessing from insights - we use the hydrated clients list from the presenter to filter for the selectedPerson by id.
updates the activeOpportunityNotifications getter to be a fn: activeNotificationsForOpportunityType so that it doesn't rely on the workflowsStore's selectedOpportunityType state field.

GitOrigin-RevId: 440c9803a3360196c6278c657488b372f6cd667d