[Supervisor Homepage] Add route and shared component for form auto-fill page (Recidiviz/recidiviz-dashboards#5989)
Adds an insights form route at /supervision/staff/:officerPseudoId/opportunity/:opportunityTypeUrl/:clientPseudoId.
The layout displayed is identical to workflows so this PR modifies the WorkflowsFormLayout component to add parameters so that the relevant data can be provided from either insights or workflows. On the insights side of things, we can re-use the opportunity presenter here.

The "back" button in the layout works as expected without any changes, however the buttons that navigate to the form needed some refactoring to be able to direct to the relevant workflows vs insights urls. The new util function getLinkToForm handles that logic.

GitOrigin-RevId: 648a5305d9ca092428c49532ec83fdf816be2808