[Supervisor Homepage] UI tweaks to metric events table (Recidiviz/recidiviz-dashboards#5906)
Finishing up the UI ticket for small changes to the metric events table for the supervisor homepage. Some margin/padding changes and hover state color change. Also it seems we made the table not scrollable for the homepage UI, which means some events were hidden so this reverts back to scrollability.

If you want to see the hover state color change, you'll need to set the supervisorHomepage feature variant for a lantern state and then navigate to that state's page. Can add something like this to your active FVs: supervisorHomepage: {activeTenants: ["US_MI"]},

GitOrigin-RevId: ba0edf28a98685a780a3860a52e039e6c973a51c