[Supervisor Homepage] P1 UI updates (methodology link, font size) (Recidiviz/recidiviz-dashboards#6067)
Two P1 tickets - updating the officer name size on the opportunity card and a methodology link that needed to open in a new tab.

Info for an officer to test who brings up the relevant methodology link is in the bug bash doc here.

Confirmed with Forrest that it is just the officer name size that needed to be increased, not the other contents of the opp card.

GitOrigin-RevId: 0e22ac9a6539a69ec1dfac73612211f892398f01