[Supervisor Homepage] Sort opportunity types by homepage position in officer presenter (Recidiviz/recidiviz-dashboards#6076)
The officer page was the only remaining spot where we needed to defer to the homepagePosition field on the opportunity config to determine what order we display opps. This PR pulls in the OppConfigurationStore to check that field and sort the list of opp types that we feed into the OpportunitySummaries component from the officer page.

GitOrigin-RevId: 2adc9cdd12612bad854a1ac3fb919ca56badd30a