[Supervisor Homepage] Fix blip on metric page by waiting for URL param to be set (Recidiviz/recidiviz-dashboards#6092)
We saw a blip when navigating from the officer page to the mertric page because the metricId field was initially undefined as it hadn't been set on the supervisionStore from the route params quite yet. This PR adds a check for the field to have been set.

It's a similar situation to the opportunity page - but it didn't feel like this warranted splitting out a whole new child presenter.

GitOrigin-RevId: fb2d1b886923e25e36295ee143f856c33706f6b5