[Workflows][Actionability] Add person-level banner to opportunity preview (Recidiviz/recidiviz-dashboards#6395)
Display a banner in the opportunity preview modal, with configurable text and a link to the selected person's profile. The banner is optional and the text is controlled via a getter on the opportunity but the logic for text to show is intended to only depend on the person. This change is behind the new FV `personSpecificOppBanners`

GitOrigin-RevId: 431a26624308e0b64d30eec78f0037ba68135fb5